refactor: avoid import from top-level ndsl within ndsl
