+++ .pre-commit-config.yaml
@@ -50,3 +50,8 @@ repos:
         args: [--autofix, --indent, "2"]
       - id: pretty-format-yaml
         args: [--autofix, --preserve-quotes, --indent, "2", --offset, "2"]
+
+  - repo: https://github.com/georgek/tach-pre-commit
+    rev: v0.32.2
+    hooks:
+      - id: tach
