+++ ndsl/dsl/dace/stree/optimizations/axis_merge.py
@@ -8,7 +8,6 @@ import dace
 import dace.sdfg.analysis.schedule_tree.treenodes as stree
 from dace.properties import CodeBlock
 
-from ndsl import ndsl_log
 from ndsl.dsl.dace.stree.optimizations.memlet_helpers import (
     AxisIterator,
     no_data_dependencies_on_cartesian_axis,
@@ -18,6 +17,7 @@ from ndsl.dsl.dace.stree.optimizations.tree_common_op import (
     list_index,
     swap_node_position_in_tree,
 )
+from ndsl.logging import ndsl_log
 
 
 # Buggy passes that should work
