+++ ndsl/dsl/dace/stree/optimizations/refine_transients.py
@@ -5,8 +5,8 @@ import warnings
 import dace.data
 import dace.sdfg.analysis.schedule_tree.treenodes as stree
 
-from ndsl import ndsl_log
 from ndsl.dsl.dace.stree.optimizations.memlet_helpers import AxisIterator
+from ndsl.logging import ndsl_log
 
 
 def _change_index_of_tuple(
