+++ ndsl/stencils/corners.py
@@ -3,9 +3,8 @@ from typing import Literal, TypeAlias, no_type_check
 from gt4py.cartesian import gtscript
 from gt4py.cartesian.gtscript import PARALLEL, computation, horizontal, interval, region
 
-from ndsl import StencilFactory
 from ndsl.constants import X_INTERFACE_DIM, Y_INTERFACE_DIM, Z_INTERFACE_DIM
-from ndsl.dsl.stencil import GridIndexing
+from ndsl.dsl.stencil import GridIndexing, StencilFactory
 from ndsl.dsl.typing import FloatField
