+++ ndsl/stencils/testing/conftest.py
@@ -8,7 +8,6 @@ import xarray as xr
 import yaml
 from f90nml import Namelist
 
-from ndsl import CompilationConfig, StencilConfig, StencilFactory
 from ndsl.comm import Comm
 from ndsl.comm.communicator import (
     Communicator,
@@ -18,6 +17,8 @@ from ndsl.comm.communicator import (
 from ndsl.comm.mpi import MPI, MPIComm
 from ndsl.comm.partitioner import CubedSpherePartitioner, TilePartitioner
 from ndsl.dsl.dace.dace_config import DaceConfig
+from ndsl.dsl.stencil import StencilFactory
+from ndsl.dsl.stencil_config import CompilationConfig, StencilConfig
 from ndsl.stencils.testing.grid import Grid  # type: ignore
 from ndsl.stencils.testing.parallel_translate import ParallelTranslate
 from ndsl.stencils.testing.savepoint import SavepointCase, Translate, dataset_to_dict
